WIP: Pool Resolver.
