+++ modules/dex-core/src/main/scala/org/ergoplatform/dex/domain/amm/package.scala
@@ -7,11 +7,14 @@ import fs2.kafka.{RecordDeserializer, RecordSerializer}
 import io.circe.{Decoder, Encoder}
 import io.estatico.newtype.macros.newtype
 import org.ergoplatform.ergo.{BoxId, TokenId}
+import sttp.tapir.{Codec, Schema, Validator}
 import tofu.logging.Loggable
 
 package object amm {
 
-  @newtype final case class PoolId(value: TokenId)
+  @newtype final case class PoolId(value: TokenId) {
+    def unwrapped: String = value.unwrapped
+  }
 
   object PoolId {
     implicit val loggable: Loggable[PoolId] = deriving
@@ -19,6 +22,14 @@ package object amm {
     // circe instances
     implicit val encoder: Encoder[PoolId] = deriving
     implicit val decoder: Decoder[PoolId] = deriving
+
+    implicit def plainCodec: Codec.PlainCodec[PoolId] = deriving
+
+    implicit def schema: Schema[PoolId] =
+      Schema.schemaForString.description("Pool ID").asInstanceOf[Schema[PoolId]]
+
+    implicit def validator: Validator[PoolId] =
+      implicitly[Validator[TokenId]].contramap[PoolId](_.value)
 
     implicit def recordSerializer[F[_]: Sync]: RecordSerializer[F, PoolId]     = serializerByEncoder
     implicit def recordDeserializer[F[_]: Sync]: RecordDeserializer[F, PoolId] = deserializerByDecoder
