+++ modules/dex-core/src/main/scala/org/ergoplatform/dex/domain/amm/state/OnChain.scala
@@ -1,9 +1,10 @@
 package org.ergoplatform.dex.domain.amm.state
 
+import derevo.circe.{decoder, encoder}
 import derevo.derive
 import tofu.logging.derivation.loggable
 
 /** On-chain entity state.
   */
-@derive(loggable)
-final case class OnChain[T](entity: T)
+@derive(loggable, encoder, decoder)
+final case class OnChain[T](onChain: T)
