+++ modules/dex-core/src/main/scala/org/ergoplatform/ergo/errors.scala
@@ -1,6 +1,10 @@
 package org.ergoplatform.ergo
 
+import tofu.Errors
+
 object errors {
 
   final case class ResponseError(msg: String) extends Exception(msg)
+
+  object ResponseError extends Errors.Companion[ResponseError]
 }
