+++ modules/pool-resolver/src/main/scala/org/ergoplatform/dex/resolver/App.scala
@@ -1,0 +1,5 @@
+package org.ergoplatform.dex.resolver
+
+object App {
+
+}
