+++ modules/pool-resolver/src/main/scala/org/ergoplatform/dex/resolver/config/ResolverConfig.scala
@@ -1,0 +1,5 @@
+package org.ergoplatform.dex.resolver.config
+
+import sttp.model.Uri
+
+final case class ResolverConfig(uri: Uri)
