+++ modules/pool-resolver/src/main/scala/org/ergoplatform/dex/resolver/repositories/Pools.scala
@@ -1,0 +1,15 @@
+package org.ergoplatform.dex.resolver.repositories
+
+import org.ergoplatform.dex.domain.amm.CFMMPool
+import org.ergoplatform.dex.domain.amm.state.{OnChain, Predicted}
+
+trait Pools[F[_]] {
+
+  /** Persist predicted pool.
+   */
+  def put(pool: Predicted[CFMMPool]): F[Unit]
+
+  /** Persist on-chain pool.
+   */
+  def put(pool: OnChain[CFMMPool]): F[Unit]
+}
