--- modules/utxo-tracker/src/main/scala/org/ergoplatform/dex/resolver/configs/TrackerConfig.scala
@@ -1,12 +1,0 @@
-package org.ergoplatform.dex.resolver.configs
-
-import derevo.derive
-import derevo.pureconfig.pureconfigReader
-import tofu.Context
-
-import scala.concurrent.duration.FiniteDuration
-
-@derive(pureconfigReader)
-final case class TrackerConfig(scanInterval: FiniteDuration, scanLastEpochs: Int, retryDelay: FiniteDuration)
-
-object TrackerConfig extends Context.Companion[TrackerConfig]
