--- modules/utxo-tracker/src/main/scala/org/ergoplatform/dex/resolver/handlers/OrdersHandler.scala
@@ -1,48 +1,0 @@
-package org.ergoplatform.dex.resolver.handlers
-
-import cats.syntax.option.none
-import cats.{Defer, Functor, FunctorFilter, Monad, MonoidK}
-import mouse.any._
-import org.ergoplatform.dex.domain.orderbook.Order.AnyOrder
-import org.ergoplatform.dex.domain.orderbook.OrderId
-import org.ergoplatform.dex.protocol.orderbook.OrderContractFamily
-import org.ergoplatform.common.streaming.{Producer, Record}
-import org.ergoplatform.dex.resolver.domain.errors.InvalidOrder
-import org.ergoplatform.dex.resolver.parsers.orders.OrdersOpsParser
-import tofu.logging.{Logging, Logs}
-import tofu.streams.{Evals, Pace}
-import tofu.syntax.handle._
-import tofu.syntax.logging._
-import tofu.syntax.monadic._
-import tofu.syntax.streams.all._
-import tofu.{Catches, Handle}
-
-final class OrdersHandler[
-  CT <: OrderContractFamily,
-  F[_]: Monad: Evals[*[_], G]: Pace: FunctorFilter: Defer: MonoidK: Catches,
-  G[_]: Monad: Handle[*[_], InvalidOrder]: Logging
-](implicit producer: Producer[OrderId, AnyOrder, F], orders: OrdersOpsParser[CT, G]) {
-
-  def handler: BoxHandler[F] =
-    _.evalMap { out =>
-      orders
-        .parseOrder(out)
-        .handleWith[InvalidOrder] { e =>
-          warnCause"Skipping invalid order in box $out" (e) as none[AnyOrder]
-        }
-    }.unNone
-      .evalTap(order => info"Order detected $order")
-      .map(o => Record[OrderId, AnyOrder](o.id, o))
-      .thrush(producer.produce)
-}
-
-object OrdersHandler {
-
-  def make[
-    CT <: OrderContractFamily,
-    I[_]: Functor,
-    F[_]: Monad: Evals[*[_], G]: Pace: FunctorFilter: Defer: MonoidK: Catches,
-    G[_]: Monad: Handle[*[_], InvalidOrder]
-  ](implicit producer: Producer[OrderId, AnyOrder, F], orders: OrdersOpsParser[CT, G], logs: Logs[I, G]): I[BoxHandler[F]] =
-    logs.forService[OrdersHandler[CT, F, G]].map(implicit log => new OrdersHandler[CT, F, G].handler)
-}
