--- modules/utxo-tracker/src/main/scala/org/ergoplatform/dex/resolver/handlers/package.scala
@@ -1,7 +1,0 @@
-package org.ergoplatform.dex.resolver
-
-import org.ergoplatform.ergo.models.Output
-
-package object handlers {
-  type BoxHandler[F[_]] = F[Output] => F[Unit]
-}
