+++ modules/utxo-tracker/src/main/scala/org/ergoplatform/dex/tracker/configs/Topics.scala
@@ -1,0 +1,8 @@
+package org.ergoplatform.dex.tracker.configs
+
+import derevo.derive
+import derevo.pureconfig.pureconfigReader
+import org.ergoplatform.common.streaming.TopicId
+
+@derive(pureconfigReader)
+final case class Topics(limitOrders: TopicId, cfmm: TopicId)
