+++ modules/utxo-tracker/src/main/scala/org/ergoplatform/dex/tracker/handlers/package.scala
@@ -1,0 +1,7 @@
+package org.ergoplatform.dex.tracker
+
+import org.ergoplatform.ergo.models.Output
+
+package object handlers {
+  type BoxHandler[F[_]] = F[Output] => F[Unit]
+}
