+++ project/versions.scala
@@ -4,14 +4,14 @@ object versions {
   val ErgoWalletVersion    = "v4.0.7-5a5356d9-SNAPSHOT"
   val ErgoContractsVersion = "1.0.0"
 
-  val SttpVersion        = "3.2.0"
-  val TapirVersion       = "0.18.0-M7"
+  val SttpVersion        = "3.3.6"
+  val TapirVersion       = "0.18.0-M17"
   val CirceVersion       = "0.13.0"
   val CatsVersion        = "2.5.0"
   val CatsEffectVersion  = "2.4.1"
   val CatsTaglessVersion = "0.12"
   val MouseVersion       = "0.26.2"
-  val Fs2Version         = "2.5.0"
+  val Fs2Version         = "2.5.6"
   val JawnFs2Version     = "1.0.0"
   val Fs2KafkaVersion    = "1.4.1"
   val JacksonVersion     = "2.12.0"
